20201203_cloudfunctions_etlsample/requirements.txt
google-cloud-bigquery==1.8.1
